@@ -3,9 +3,9 @@
   "version": "0.8.15",
   "description": "An official implementation of the Level Up: Advanced 5th Edition system for FoundryVTT.",
   "main": "dist/index.js",
+  "type": "module",
   "scripts": {
     "dev": "vite",
-
     "build": "npm run build:compendia && vite build",
     "build:system": "vite build",
     "build:compendia": "node -e \"import('./compendia.mjs').then( m => m.compilePacks())\"",
@@ -32,30 +32,25 @@
   "license": "MIT",
   "devDependencies": {
     "@typhonjs-fvtt/eslint-config-foundry.js": "^0.8.0",
-    "@vue/compiler-sfc": "^3.2.37",
-    "eslint": "^8.20.0",
+    "eslint": "^8.35.0",
     "eslint-config-airbnb": "^19.0.4",
-    "eslint-plugin-import": "^2.26.0",
-    "eslint-plugin-jsx-a11y": "^6.6.1",
-    "eslint-plugin-react": "^7.30.1",
-    "glob": "^8.0.3",
-    "node-sass": "^7.0.1",
-    "postcss-scss": "^4.0.4",
-    "sass": "^1.54.0",
-    "stylelint": "^14.9.1",
+    "eslint-plugin-import": "^2.27.5",
+    "eslint-plugin-jsx-a11y": "^6.7.1",
+    "eslint-plugin-react": "^7.32.2",
+    "eslint-plugin-svelte3": "^4.0.0",
+    "glob": "^9.2.1",
+    "postcss-scss": "^4.0.6",
+    "sass": "^1.58.3",
+    "stylelint": "^14.16.1",
     "stylelint-config-sass-guidelines": "^9.0.1",
-    "vite": "^2.9.14"
+    "vite": "^4.1.4"
   },
   "dependencies": {
-    "@tiptap/extension-image": "^2.0.0-beta.27",
-    "@tiptap/extension-link": "^2.0.0-beta.37",
-    "@tiptap/extension-text-align": "^2.0.0-beta.29",
-    "@tiptap/starter-kit": "^2.0.0-beta.184",
-    "@tiptap/vue-3": "^2.0.0-beta.91",
-    "@vitejs/plugin-vue": "^2.3.3",
-    "handlebars-helper-range": "^1.1.1",
-    "remixicon": "^2.5.0",
-    "vue": "^3.2.37",
-    "vue-input-autowidth": "^2.2.0"
+    "@rollup/plugin-node-resolve": "^15.0.1",
+    "@sveltejs/vite-plugin-svelte": "^2.0.3",
+    "@typhonjs-fvtt/runtime": "^0.0.22",
+    "@typhonjs-fvtt/svelte-standard": "^0.0.19",
+    "svelte": "^3.55.1",
+    "svelte-preprocess": "^5.0.0"
   }
 }
