@@ -1,6 +1,6 @@
-import vue from '@vitejs/plugin-vue';
-
-const path = require('path');
+import { svelte } from '@sveltejs/vite-plugin-svelte';
+import preprocess from 'svelte-preprocess';
+import path from 'path';
 
 const config = {
   root: 'src/',
@@ -29,7 +29,22 @@ const config = {
       fileName: () => 'index.js'
     }
   },
-  plugins: [vue()]
+  plugins: [
+    svelte({
+      preprocess: preprocess(),
+      onwarn: (warning, handler) => {
+        // Suppress `a11y-missing-attribute` for missing href in <a> links.
+        // Foundry doesn't follow accessibility rules.
+        if (warning.message.includes('<a> element should have an href attribute')) return;
+        if (warning.code === 'a11y-click-events-have-key-events') return;
+
+        console.log(warning);
+
+        // Let Rollup handle all other warnings normally.
+        handler(warning);
+      }
+    })
+  ]
 };
 
 export default config;
